# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.PaymentIntentRequest:
    properties:
      payment_method_id:
        type: integer
    type: object
...
  model.UpdatePaymentIntentRequestBody:
    properties:
      failur_reason:
        type: string
      status:
        enum:
        - SUCCESS
        - FAILED
        example: SUCCESS
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.ErrorBody'
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for Project VioPal
  title: VioPal
  version: "1.0"
paths:
...
  /api/v1/admin/payment-intents:
    get:
      consumes:
      - application/json
      description: Get payment intent list with pagination and optional filters
      parameters:
      - description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - description: Items per page
        example: 10
        in: query
        name: perpage
        type: integer
      - description: Payment intent status
        enum:
        - PENDING
        - SUCCESS
        - FAILED
        - EXPIRED
        in: query
        name: status
        type: string
      - description: Payment intent number
        in: query
        name: payment_intent_number
        type: string
      - description: Payment method ID
        in: query
        name: payment_method_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List payment intents
      tags:
      - Admin
  /api/v1/admin/payment-intents/{payment_intent_number}/status:
    patch:
      consumes:
      - application/json
      description: Update payment intent status by payment intent number (Admin only).
        If status is FAILED, failure_reason is required.
      parameters:
      - description: Payment Intent Number
        in: path
        name: payment_intent_number
        required: true
        type: string
      - description: Update Payment Intent Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePaymentIntentRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update payment intent status
      tags:
      - Admin
...
  /api/v1/public/intents/{code}:
    get:
      description: Get payment intent detail using public intent code
      parameters:
      - description: Payment Intent Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get payment intent by code
      tags:
      - Payment
  /api/v1/public/pay/{token}:
    get:
      description: Process payment using payment link token
      parameters:
      - description: Payment Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Process payment by token
      tags:
      - Payment
  /api/v1/public/pay/{token}/intents:
    post:
      consumes:
      - application/json
      description: Create payment intent using payment link token and selected payment
        method
      parameters:
      - description: Payment Token
        in: path
        name: token
        required: true
        type: string
      - description: Payment Method Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PaymentIntentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create payment intent
      tags:
      - Payment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"